+++ ksa_compliance/public/js/sales_invoice.js
@@ -31,17 +31,57 @@ frappe.require("/assets/ksa_compliance/js/update_invoice_mode_of_payment.js").th
             frm.trigger("apply_advance_payments");
         },
         mode_of_payment: function(frm){
-            frappe.db.get_value(
-                "Mode of Payment",
-                { mode_of_payment: frm.doc.mode_of_payment },
-                "type"
-            ).then(response => {
-                const type = response.message.type;
-                if (type === "Bank"){
-                    frm.set_df_property("reference_no", "read_only", 0);
-                    frm.toggle_reqd("reference_no", 1);
-                    frm.set_df_property("reference_date", "read_only", 0);
-                    frm.toggle_reqd("reference_date", 1);
+            frm.trigger("fetch_mode_of_payment_account");
+        },
+        mode_of_payment_account: function (frm){
+            frm.trigger("setup_reference_no_and_date_properties");
+        },
+        fetch_mode_of_payment_account: function (frm){
+            frappe.call({
+                method: "erpnext.accounts.doctype.sales_invoice.sales_invoice.get_bank_cash_account",
+                args: {
+                    mode_of_payment: frm.doc.mode_of_payment,
+					company: frm.doc.company,
+                },
+                callback: function(r) {
+                    if(r.message) {
+                        frm.set_value("mode_of_payment_account", r.message.account);
+                    }
+                }
+            });
+            frm.set_df_property("mode_of_payment_account", "read_only", 0);
+            frm.refresh_field("mode_of_payment_account");
+            frm.set_query('mode_of_payment_account', function (doc){
+                return {
+                    filters: [
+                        ["Account", "account_type", "in", "Bank, Cash, Receivable"],
+                        ["Account", "is_group", "=", 0],
+                        ["Account", "company", "=", frm.doc.company],
+                    ],
+                };
+            });
+        },
+        setup_reference_no_and_date_properties: function (frm, mode_of_payment_account){
+            frappe.call({
+                method: "erpnext.accounts.doctype.payment_entry.payment_entry.get_account_details",
+                args: {
+                    account:frm.doc.mode_of_payment_account,
+                    date:frm.doc.posting_date,
+                    cost_center:frm.doc.cost_center,
+                },
+                callback: function(response) {
+                    const type = response.message.account_type;
+                    if (type === "Bank"){
+                        frm.set_df_property("reference_no", "read_only", 0);
+                        frm.toggle_reqd("reference_no", 1);
+                        frm.set_df_property("reference_date", "read_only", 0);
+                        frm.toggle_reqd("reference_date", 1);
+                    }else {
+                        frm.set_df_property("reference_no", "read_only", 1);
+                        frm.toggle_reqd("reference_no", 0);
+                        frm.set_df_property("reference_date", "read_only", 1);
+                        frm.toggle_reqd("reference_date", 0);
+                    }
                     frm.refresh_field("reference_no");
                     frm.refresh_field("reference_date");
                 }